wip: testing: Make tcpdump write output line buffered and packet buffered
With newer versions we don't seem to get any packets in the file
otherwise.
WIP: actually doesn't fix the issue yet. What we get when terminating
tcpdump is this:

0 packets captured
18 packets received by filter
0 packets dropped by kernel

So 18 packets were captured by libpcap but tcpdump did not yet process
and print them. Writing to a capture file does not seem to help.
This might be due to the timeout used in the poll() calls in libpcap and
set via pcap_set_timeout() by tcpdump or some buffering. We might have to
provide our own patched tcpdump/libpcap build.