wip: testing: Install newer tcpdump from jessie-backports, which supports --immediate-mode
Without this we get something like the following when terminating tcpdump
after the short running time of a scenario:

0 packets captured
18 packets received by filter
0 packets dropped by kernel

So 18 packets were captured by libpcap but tcpdump did not yet process
and print them.  With immediate mode enabled there is no buffering in
libpcap so tcpdump gets the packets instantly and we get the expected
output.

The version in wheezy still works, but it does not support
--immediate-mode, so we check whether tcpdump supports it or not
before running the tests.  There is no working backport for wheezy that
would support the option either (the one in wheezy-backports-sloppy has
a dependency on a new version of libc that can't be resolved).

WIP: While this works pretty well, there are sometimes still situations
like the one above. And in some cases we don't even see the final messages
in the log, just the first two status lines.