WIP: testing: Use TLS 1.2 in RADIUS test cases
This requires a patch in the OpenSSL package.  In the release currently
shipped with Debian wheezy and on which our FIPS-enabled package is based,
the function SSL_export_keying_material(), which is used by FreeRADIUS to
derive the MSK, does not use the correct digest to calculate the result
when TLS 1.2 is used.  This causes IKE to fail with "verification of AUTH
payload with EAP MSK failed".  This can be fixed by passing the result
of ssl_get_algorithm2(s) to tls1_PRF() instead of directly using the value
s->s3->tmp.new_cipher->algorithm2.