wip: getaddrinfo behavior changed
Another issue that came up is a change in the behavior of getaddrinfo.
We don't have a resolv.conf in most test scenarios, still getaddrinfo()
does seem to attempt a DNS lookup for unknown host names via localhost,
which times out after about 20s (or 10s when done by curl where we set that
timeout specifically).  In earlier releases such a lookup was not done and
the resolution of a non-existing hostnames failed pretty much immediately.

Changing hosts in nsswitch.conf from "files dns" to just "files" helps, so
we might just do that and fix nsswitch.conf for the few scenarios where
we actually do want to do DNS resolution (dnssec related).